Switching to custom server
